refactor: move tests/ to evaluator/test_classes/
Move all test class definitions and evaluator unit tests from the
top-level tests/ directory into evaluator/test_classes/, keeping
them co-located with the evaluator module that depends on them.

- 17 files renamed: tests/* -> evaluator/test_classes/*
- Updated imports in evaluator/engine.py, scoring.py, and
  strategies/two_pass.py to use the new module path
- Updated 9 internal test imports in test_answer_extractor.py
- No logic changes; pure relocation
